JPD_Compiler를 통해 생성된 RPC 코드 반영 및 테스트

+++ RPC_orgn/FightGameCrtDel.cs
@@ -1,40 +1,38 @@
-
 using System;
+using System.Collections;
+using System.Collections.Generic;
 using UnityEngine;
 
 public class FightGameCrtDel : Stub_FightGameCrtDel
 {
-    private void Start() 
+    private void Start()
     {
         base.Init();
     }
 
-
-    protected override void CRT_CHARACTER(UInt32 ID, Byte Dir, UInt16 X, UInt16 Y, Byte HP) 
+    protected override void CRT_CHARACTER(UInt32 ID, byte Dir, UInt16 X, UInt16 Y, byte HP)
     {
         BattleField battleField = gameObject.GetComponent<BattleField>();
-        Fighter fighter = battleField.CreateFighter(ID, Dir, X, Y, HP);
-        if (fighter.gameObject.GetComponent<FighterController>() == null)
+        Fighter fighter = battleField.CreateFighter(ID, Dir, X, Y, HP);    
+        if(fighter.gameObject.GetComponent<FighterController>() == null)
         {
-            fighter.gameObject.AddComponent<FighterController>();
+            fighter.gameObject.AddComponent<FighterController>();   
         }
 
         GameObject mainCamera = GameObject.Find("Main Camera");
         mainCamera.transform.parent = fighter.gameObject.transform;
         mainCamera.transform.localPosition = new Vector3(0, 0, -10);
 
-        battleField.Player = fighter;
+        battleField.Player = fighter;   
         battleField.Fighters.Add(ID, fighter);
     }
-
-    protected override void CRT_OTHER_CHARACTER(UInt32 ID, Byte Dir, UInt16 X, UInt16 Y, Byte HP) 
+    protected override void CRT_OTHER_CHARACTER(UInt32 ID, byte Dir, UInt16 X, UInt16 Y, byte HP)
     {
         BattleField battleField = gameObject.GetComponent<BattleField>();
         Fighter fighter = battleField.CreateFighter(ID, Dir, X, Y, HP);
         battleField.Fighters.Add(ID, fighter);
     }
-
-    protected override void DEL_CHARACTER(UInt16 ID) 
+    protected override void DEL_CHARACTER(UInt32 ID)
     {
         BattleField battleField = gameObject.GetComponent<BattleField>();
         if (battleField.Fighters.ContainsKey(ID))
@@ -56,5 +54,4 @@ public class FightGameCrtDel : Stub_FightGameCrtDel
             }
         }
     }
-
 }
